# ch11/secret-diary/docs/4-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "나만의 비밀일기 API",
    "description": "개인 인증 기반 프라이빗 일기 서비스 REST API. 자체 JWT(Access 12h / Refresh 7d) 인증 사용.\n\n관련 문서: docs/1-prd.md, docs/2-erd.md, docs/3-supabase-ddl.sql",
    "version": "1.0.0",
    "contact": { "name": "Secret Diary Team" }
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "로컬 개발" }
  ],
  "tags": [
    { "name": "Auth", "description": "회원가입/로그인/토큰 (FR-1)" },
    { "name": "Diaries", "description": "일기 CRUD 및 필터 (FR-2)" },
    { "name": "Users", "description": "사용자 정보/비밀번호 (FR-3)" }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "회원 가입 (FR-1.1)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "가입 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfile" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "email 또는 username 중복",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "code": "CONFLICT", "message": "이미 사용 중인 이메일입니다." }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인 (FR-1.3, FR-1.4)",
        "description": "email 또는 username + password 로 로그인. 성공 시 Access/Refresh Token 발급.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": {
            "description": "계정 없음 또는 비밀번호 불일치 (정보 노출 방지 위해 동일 메시지)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "code": "UNAUTHORIZED", "message": "이메일 또는 비밀번호가 올바르지 않습니다." }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Access Token 재발급 (FR-1.5)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string", "description": "발급받은 Refresh Token" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "재발급 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "401": {
            "description": "Refresh Token 만료/폐기/위조",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그아웃 - Refresh Token 무효화 (FR-1.6)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "로그아웃 완료" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/diaries": {
      "get": {
        "tags": ["Diaries"],
        "summary": "내 일기 목록 조회 + 필터 (FR-2.3, FR-2.7)",
        "description": "본인 일기만 diaryDate 최신순(동일 날짜는 createdAt 순)으로 조회. weather/mood/tag 로 복합 필터링 가능.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "weather", "in": "query", "required": false,
            "schema": { "$ref": "#/components/schemas/Weather" },
            "description": "날씨 필터"
          },
          {
            "name": "mood", "in": "query", "required": false,
            "schema": { "$ref": "#/components/schemas/Mood" },
            "description": "기분 필터"
          },
          {
            "name": "tag", "in": "query", "required": false,
            "schema": { "type": "string" },
            "description": "태그명 필터 (단일)"
          },
          {
            "name": "page", "in": "query", "required": false,
            "schema": { "type": "integer", "minimum": 1, "default": 1 }
          },
          {
            "name": "limit", "in": "query", "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공 (결과 0건이면 빈 배열)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DiaryListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Diaries"],
        "summary": "일기 작성 (FR-2.1)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DiaryCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "작성 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Diary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/diaries/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "tags": ["Diaries"],
        "summary": "일기 상세 조회 (FR-2.4)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Diary" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Diaries"],
        "summary": "일기 수정 (FR-2.5)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DiaryCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Diary" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Diaries"],
        "summary": "일기 삭제 (FR-2.6)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "삭제 완료" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "내 프로필 + 작성 일기 수 조회 (FR-3.1, FR-3.2)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserProfileWithStats" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/users/me/password": {
      "put": {
        "tags": ["Users"],
        "summary": "비밀번호 변경 (FR-3.3)",
        "description": "현재 비밀번호 확인 후 변경.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "변경 완료" },
          "400": {
            "description": "현재 비밀번호 불일치 또는 새 비밀번호 정책 미달",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token (유효 12시간). Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "Weather": {
        "type": "string",
        "enum": ["sunny", "cloudy", "rainy", "snowy", "windy"],
        "description": "날씨: 맑음/흐림/비/눈/바람"
      },
      "Mood": {
        "type": "string",
        "enum": ["happy", "neutral", "sad", "angry", "excited", "tired"],
        "description": "기분: 행복/보통/슬픔/화남/설렘/피곤"
      },
      "SignupRequest": {
        "type": "object",
        "required": ["email", "username", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "username": { "type": "string", "minLength": 2, "maxLength": 50, "example": "jimin" },
          "password": { "type": "string", "format": "password", "minLength": 8, "example": "P@ssw0rd!" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["identifier", "password"],
        "properties": {
          "identifier": { "type": "string", "description": "email 또는 username", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "P@ssw0rd!" }
        }
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string", "description": "JWT Access Token (12시간)" },
          "refreshToken": { "type": "string", "description": "Refresh Token (7일)" },
          "tokenType": { "type": "string", "example": "Bearer" },
          "expiresIn": { "type": "integer", "description": "Access Token 만료(초)", "example": 43200 }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserProfileWithStats": {
        "allOf": [
          { "$ref": "#/components/schemas/UserProfile" },
          {
            "type": "object",
            "properties": {
              "diaryCount": { "type": "integer", "description": "작성한 일기 총 개수 (FR-3.2)", "example": 42 }
            }
          }
        ]
      },
      "PasswordChangeRequest": {
        "type": "object",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string", "format": "password" },
          "newPassword": { "type": "string", "format": "password", "minLength": 8 }
        }
      },
      "DiaryCreateRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200, "example": "비 오는 날의 기록" },
          "content": { "type": "string", "minLength": 1, "example": "오늘은 하루 종일 비가 내렸다." },
          "weather": { "$ref": "#/components/schemas/Weather" },
          "mood": { "$ref": "#/components/schemas/Mood" },
          "diaryDate": {
            "type": "string",
            "format": "date",
            "description": "일기가 다루는 날짜(YYYY-MM-DD). 미입력 시 오늘 날짜로 설정된다. 지난 날짜도 지정 가능(예: 어제 못 쓴 일기를 오늘 작성)",
            "example": "2026-07-05"
          },
          "tags": {
            "type": "array",
            "items": { "type": "string", "maxLength": 20 },
            "maxItems": 10,
            "description": "태그명 배열. 최대 10개, 각 20자 (PRD 8)",
            "example": ["일상", "비"]
          }
        }
      },
      "Diary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "weather": { "$ref": "#/components/schemas/Weather" },
          "mood": { "$ref": "#/components/schemas/Mood" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "diaryDate": { "type": "string", "format": "date", "description": "일기가 다루는 날짜(목록 정렬 기준)" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DiaryListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Diary" } },
          "pagination": {
            "type": "object",
            "properties": {
              "page": { "type": "integer", "example": 1 },
              "limit": { "type": "integer", "example": 20 },
              "total": { "type": "integer", "example": 42 },
              "totalPages": { "type": "integer", "example": 3 }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "BAD_REQUEST" },
          "message": { "type": "string", "example": "요청이 올바르지 않습니다." },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            },
            "description": "필드별 검증 오류 (선택)"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청 / 유효성 검증 실패",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "인증 실패 (토큰 없음/만료/위조)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "code": "UNAUTHORIZED", "message": "인증이 필요합니다." }
          }
        }
      },
      "Forbidden": {
        "description": "타인 리소스 접근 (소유권 없음)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "code": "FORBIDDEN", "message": "접근 권한이 없습니다." }
          }
        }
      },
      "NotFound": {
        "description": "리소스 없음",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "code": "NOT_FOUND", "message": "일기를 찾을 수 없습니다." }
          }
        }
      }
    }
  }
}